[Core] Move hash_request_tokens computation from input request threads
