+++ vllm/v1/core/kv_cache_manager.py
@@ -7,13 +7,12 @@ from typing import Optional
 
 from vllm.distributed.kv_events import KVCacheEvent
 from vllm.logger import init_logger
-from vllm.utils import sha256, sha256_cbor_64bit
 from vllm.v1.core.kv_cache_coordinator import get_kv_cache_coordinator
 from vllm.v1.core.kv_cache_utils import (BlockHash, KVCacheBlock,
-                                         hash_request_tokens, init_none_hash)
+                                         get_hash_fn_by_name)
 from vllm.v1.kv_cache_interface import KVCacheConfig
 from vllm.v1.metrics.stats import PrefixCacheStats
-from vllm.v1.request import Request, RequestStatus
+from vllm.v1.request import Request, RequestStatus, hash_request_tokens
 
 logger = init_logger(__name__)
 
@@ -84,10 +83,7 @@ class KVCacheManager:
             enable_caching = False
         self.enable_caching = enable_caching
 
-        self.caching_hash_fn = (
-            sha256_cbor_64bit if caching_hash_algo == "sha256_cbor_64bit" else
-            sha256 if caching_hash_algo == "sha256" else hash)
-        init_none_hash(self.caching_hash_fn)
+        self.caching_hash_fn = get_hash_fn_by_name(caching_hash_algo)
         self.use_eagle = use_eagle
         self.log_stats = log_stats
         # FIXME: make prefix cache stats conditional on log_stats
@@ -166,8 +162,12 @@ class KVCacheManager:
         block_hashes = self.req_to_block_hashes[request.request_id]
         if not block_hashes:
             assert self.block_size is not None
-            block_hashes = hash_request_tokens(self.caching_hash_fn,
-                                               self.block_size, request)
+            if request.precomputed_block_hashes is not None:
+                block_hashes = request.precomputed_block_hashes
+            else:
+                block_hashes = hash_request_tokens(self.caching_hash_fn,
+                                                   self.block_size, request)
+
             self.req_to_block_hashes[request.request_id] = block_hashes
 
         # NOTE: When all tokens hit the cache, we must recompute the last token
